+++ src/modules/contacts/adapter/gateways/ContactsRepositoryFactory.ts
@@ -1,0 +1,8 @@
+import { FirestoreContactsRepository } from "./FirestoreContactsRepository";
+import { firestore } from "../../../../frameworks/firestore/firestore";
+
+export class ContactsRepositoryFactory {
+  static getRepository() {
+    return new FirestoreContactsRepository(firestore);
+  }
+}