+++ src/modules/contacts/entities/Contact.ts
@@ -1,21 +1,26 @@
 import { IDType } from "../../shared/entities/types";
-import { IPerson } from "./IPerson";
+import { IContact } from "./IContact";
 
-export class Person implements IPerson {
+export class Contact implements IContact {
   id: IDType;
   name: string;
-  createdBy: IDType;
+  email: string;
+  country: string;
   createdAt: string;
   updatedAt: string;
 
   constructor(values: {
     id: IDType,
     name: string,
+    email: string,
+    country: string,
     createdAt: string,
     updatedAt: string,
   }) {
     this.id = values.id;
     this.name = values.name;
+    this.email = values.name;
+    this.country = values.name;
     this.createdAt = values.createdAt;
     this.updatedAt = values.updatedAt;
   }