+++ src/modules/contacts/useCases/sharedPorts/IContactsRepository.ts
@@ -1,0 +1,10 @@
+import { IContact } from "../../entities/IContact";
+import { IRepositories } from "../../../shared/useCases/ports/IRepositories";
+
+export interface IContactsRepository extends IRepositories<IContact> {
+  create(values: {
+    name: string;
+    email: string;
+    country: string;
+  }): Promise<IContact>;
+}