+++ src/modules/people/adapter/middlewares/createPersonBody.ts
@@ -1,7 +1,0 @@
-import Joi from "joi";
-
-export const createPersonBody = Joi.object({
-  name: Joi.string()
-    .min(1)
-    .required(),
-});